Emulate 6280A to eliminate 'pop' sound when PSG channel is activated
The issue was caused by using a '0' default value - instead of mid-range - for any PSG channel which is inactive; activating it causes an apparent DC bias which causes the pop sound.  The solution is to use a midrange value as the default for inactive channels instead.

HOWEVER: this replicates the 6280A CPU which was used only in SuperGrafx and some CoreGrafx machines.  Additional effort would be required to emulate both chips.  At this time, I am not aware of any games or situations where the staccato nature of the original 6280 would be preferable.  I will only implement it if there is such a case.